requirements.txt
git+https://github.com/Maayanlab/appyter.git@455c8dc9c761aea4482afed1c5a405ced9223b84
pandas
numpy
s3fs
git+https://github.com/Maayanlab/maayanlab-bioinformatics.git
igraph
matplotlib_venn
leidenalg
fsspec==2022.1.0
tqdm
scanpy
rpy2==3.5.10
fastcluster
openai
python-dotenv
networkx
biopython
